Implemented the new read factory paradigm
All read factory classes must have a method named getReadNamesToLoad which
generates a list of names of reads which may be loaded from this source
and which satisfy the criteria passed to the constructor (e.g. asped date
constraints).

All read factory classes must also have a method named getReadByName
which takes a read name and constructs a Read object for that name from
data taken from the source.


git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/arcturus/trunk@1760 69345dfe-1762-0410-a7d4-b9dc18918749